backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
oracledb>=2.0.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
pydantic[email]==2.10.4
python-multipart==0.0.18
boto3==1.37.38
Pillow>=10.0.0
google-genai>=1.0.0
python-telegram-bot>=21.0
